codefactor.io feedback again

+++ deploy/deploy.sh
@@ -15,14 +15,14 @@ rm -r $TARGET_DIR
 
 echo "Empty $TMP_DIR"
 mkdir -p $TMP_DIR
-rm -r $TMP_DIR/*
+rm -r "${TMP_DIR:?}/"*
 
 echo 'Install npm dependencies'
 cd $SRC_DIR
 for d in $(ls -1)
 do
   cd $d
-  if [ -z package.json ]
+  if [ -z "package.json" ]
   then
     npm install > $TMP_DIR/$d_npm_install.log
   fi
